Make error handling robust
