+++ posthog/hogql/test/test_parser_python.py
@@ -1,5 +1,5 @@
 from ._test_parser import parser_test_factory
 
 
-class TestParserCpp(parser_test_factory("cpp")):
+class TestParserPython(parser_test_factory("python")):
     pass
